+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - added unobtainable secrets stat support in the gameflow (#1379)
 - added a wireframe mode
 - changed console caret blinking rate (#1377)
+- changed the TR1X install source in the installer to suggest using the existing installation directory (#1350)
 - fixed config tool and installer missing icons (#1358, regression from 4.0)
 - fixed looking forward too far causing an upside down camera frame (#1338)
 - fixed the enemy bear behavior in demo mode (#1370, regression since 2.16)